LayoutTests/imported/w3c/web-platform-tests/css/cssom/getComputedStyle-pseudo-expected.txt


PASS Resolution of width is correct for ::before and ::after pseudo-elements 
PASS Resolution of width is correct for ::before and ::after pseudo-elements of display: contents elements 
PASS Resolution of nonexistent pseudo-element styles 
FAIL Resolution of pseudo-element styles in display: none elements assert_equals: Pseudo-styles of display: none elements should be correct expected "\"Foo\"" but got "Foo"
PASS Item-based blockification of pseudo-elements 
FAIL Item-based blockification of nonexistent pseudo-elements assert_equals: Pseudo-styles of display: flex elements should get blockified expected "block" but got "inline"
PASS display: contents on pseudo-elements 
FAIL Dynamically change to display: contents on pseudo-elements assert_equals: display: contents in :before should get reflected on CSSOM expected "contents" but got "block"
FAIL Unknown pseudo-elements throw assert_throws_js: getComputedStyle with an unknown pseudo-element throws function "() => getComputedStyle(div, "totallynotapseudo")" did not throw
